requirements.txt
discord.py>=1.7.2
discord-py-slash-command==1.2.0
pytz